+++ rasa_core/run.py
@@ -12,6 +12,7 @@ from rasa_core import utils
 from rasa_core.agent import Agent
 from rasa_core.channels.console import ConsoleInputChannel
 from rasa_core.channels.facebook import FacebookInput
+from rasa_core.channels.telegram import TelegramInput
 from rasa_core.channels.rest import HttpInputChannel
 from rasa_core.channels.slack import SlackInput
 from rasa_core.utils import read_yaml_file
@@ -51,50 +52,35 @@ def create_argument_parser():
     parser.add_argument(
         '-c', '--connector',
         default="cmdline",
-        choices=["facebook", "slack", "cmdline"],
+        choices=["facebook", "slack", "telegram", "cmdline"],
         help="service to connect to")
 
-    # arguments for logging configuration
-    parser.add_argument(
-        '--debug',
-        help="Print lots of debugging statements. "
-             "Sets logging level to DEBUG",
-        action="store_const",
-        dest="loglevel",
-        const=logging.DEBUG,
-        default=logging.WARNING,
-    )
-    parser.add_argument(
-        '-v', '--verbose',
-        help="Be verbose. Sets logging level to INFO",
-        action="store_const",
-        dest="loglevel",
-        const=logging.INFO,
-    )
-
+    utils.add_logging_option_arguments(parser)
     return parser
 
 
-def _raise_missing_credentials_Exception(channel):
+def _raise_missing_credentials_exception(channel):
     if channel == "facebook":
         channel_doc_link = "facebook-messenger"
     elif channel == "slack":
         channel_doc_link = "slack"
+    elif channel == "telegram":
+        channel_doc_link = "telegram"
     else:
         channel_doc_link = ""
 
-    raise Exception("To use the slack input channel, you need to "
+    raise Exception("To use the {} input channel, you need to "
                     "pass a credentials file using '--credentials'. "
                     "The argument should be a file path pointing to"
                     "a yml file containing the {} authentication"
                     "information. Details in the docs: "
                     "https://core.rasa.ai/connectors.html#{}-setup".
-                    format(channel, channel_doc_link))
+                    format(channel, channel, channel_doc_link))
 
 
 def _create_external_channel(channel, port, credentials_file):
     if credentials_file is None:
-        _raise_missing_credentials_Exception(channel)
+        _raise_missing_credentials_exception(channel)
 
     credentials = read_yaml_file(credentials_file)
     if channel == "facebook":
@@ -106,9 +92,14 @@ def _create_external_channel(channel, port, credentials_file):
         input_blueprint = SlackInput(
             credentials.get("slack_token"),
             credentials.get("slack_channel"))
+    elif channel == "telegram":
+        input_blueprint = TelegramInput(
+            credentials.get("access_token"),
+            credentials.get("verify"),
+            credentials.get("webhook_url"))
     else:
-        Exception("This script currently only supports the facebook "
-                  "and slack connectors.")
+        Exception("This script currently only supports the facebook,"
+                  " telegram and slack connectors.")
 
     return HttpInputChannel(port, None, input_blueprint)
 
@@ -116,7 +107,7 @@ def _create_external_channel(channel, port, credentials_file):
 def create_input_channel(channel, port, credentials_file):
     """Instantiate the chosen input channel."""
 
-    if channel == "facebook" or channel == "slack":
+    if channel in ['facebook', 'slack', 'telegram']:
         return _create_external_channel(channel, port, credentials_file)
     elif channel == "cmdline":
         return ConsoleInputChannel()