+++ tests/test_tracker_stores.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 
 from rasa_core.channels import UserMessage
 from rasa_core.domain import TemplateDomain
-from rasa_core.events import SlotSet
+from rasa_core.events import SlotSet, ActionExecuted, Restarted
 from rasa_core.tracker_store import InMemoryTrackerStore
 
 domain = TemplateDomain.load("data/test_domains/default_with_topic.yml")
@@ -25,3 +25,20 @@ def test_get_or_create():
 
     again = store.get_or_create_tracker(UserMessage.DEFAULT_SENDER_ID)
     assert again.get_slot(slot_key) == slot_val
+
+
+def test_restart_after_retrieval_from_tracker_store(default_domain):
+    store = InMemoryTrackerStore(default_domain)
+    tr = store.get_or_create_tracker("myuser")
+    synth = [ActionExecuted("action_listen") for _ in range(4)]
+
+    for e in synth:
+        tr.update(e)
+
+    tr.update(Restarted())
+    latest_restart = tr.idx_after_latest_restart()
+
+    store.save(tr)
+    tr2 = store.retrieve("myuser")
+    latest_restart_after_loading = tr2.idx_after_latest_restart()
+    assert latest_restart == latest_restart_after_loading