+++ tests/test_training.py
@@ -26,6 +26,21 @@ def test_story_visualization(default_domain):
                                         should_merge_nodes=False)
     assert len(generated_graph.nodes()) == 33
 
+    assert len(generated_graph.edges()) == 36
+
+
+def test_story_visualization_with_merging(default_domain):
+    story_steps = StoryFileReader.read_from_file(
+            "data/test_stories/stories.md", default_domain,
+            interpreter=RegexInterpreter())
+    generated_graph = visualize_stories(story_steps, default_domain,
+                                        output_file=None,
+                                        max_history=3,
+                                        should_merge_nodes=True)
+    assert 15 < len(generated_graph.nodes()) < 33
+
+    assert 20 < len(generated_graph.edges()) < 33
+
 
 def test_training_script(tmpdir):
     train_dialogue_model(DEFAULT_DOMAIN_PATH, DEFAULT_STORIES_FILE,
@@ -33,3 +48,4 @@ def test_training_script(tmpdir):
                          use_online_learning=False,
                          nlu_model_path=None,
                          kwargs={})
+    assert True