+++ CHANGELOG.rst
@@ -4,26 +4,68 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning`_ starting with version 0.2.0.
 
-[Unreleased] - `master`_
-^^^^^^^^^^^^^^^^^^^^^^^^
+[Unreleased 0.9.0.aX] - `master`_
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. note:: This version is not yet released and is under active development.
 
-This is a major version change. Make sure to take a look at the migrations
-guide in the documentation for advice on how to update existing projects.
+Added
+-----
+
+Changed
+-------
+
+Removed
+-------
+
+Fixed
+-----
+
+[0.8.2] - 2018-02-13
+^^^^^^^^^^^^^^^^^^^^
+
+Added
+-----
+- script to reload a dumped trackers state and to continue the conversation
+  at the end of the stored dialogue
+
+Changed
+-------
+- minor updates to dependencies
+
+Fixed
+-----
+- fixed datetime serialisation of reminder event
+
+[0.8.1] - 2018-02-01
+^^^^^^^^^^^^^^^^^^^^
+
+Fixed
+-----
+- removed deque to support python 3.5
+- Documentation improvements to tutorials
+- serialisation of date time value for ``ReminderScheduled`` event
+
+[0.8.0] - 2018-01-30
+^^^^^^^^^^^^^^^^^^^^
+
+This is a major version change. Make sure to take a look at the
+:ref:`migration` in the documentation for advice on how to
+update existing projects.
 
 Added
 -----
 - ``--debug`` and ``--verbose`` flags to scripts (train.py, run.py, server.py)
   to set the log level
 - support for story cycles when using checkpoints
-- added a new machine learning policy `SklearnPolicy` that uses an sklearn 
+- added a new machine learning policy `SklearnPolicy` that uses an sklearn
   classifier to predict actions (logistic regression by default)
 - warn if action emits events when using a model that it did never emit in
   any of the stories the model was trained on
 - support for event pushing and endpoints to retrieve the tracker state from the server
 - Timestamp to every event
 - added a Slack channel that allows Rasa Core to communicate via a Slack app
+- added a Telegram channel that allows Rasa Core to communicate via a Telegram bot
 
 Changed
 -------
@@ -35,11 +77,8 @@ Changed
 - improved travis build speed by not using miniconda
 - don't fail with an exception but with a helpful error message if an
   utterance template contains a variable that can not be filled
-- domain doesn't fail on unknown actions but emits a warning instead. this is to support reading 
+- domain doesn't fail on unknown actions but emits a warning instead. this is to support reading
   logs from older conversation if one recently removed an action from the domain
-
-Removed
--------
 
 Fixed
 -----
@@ -48,6 +87,10 @@ Fixed
 - fixed float slot min max value handling
 - fixed non integer feature decoding, e.g. used for memoization policy
 - properly log to specified file when starting Rasa Core server
+- properly calculate offset of last reset event after loading tracker from
+  tracker store
+- UserUtteranceReverted action incorrectly triggered actions to be replayed
+
 
 [0.7.9] - 2017-11-29
 ^^^^^^^^^^^^^^^^^^^^